fix: uploadtile renamed to uploadstatus
