+++ dataprep-webapp/src/app/components/dataset/upload-status/_dataset-upload-status.scss
@@ -11,7 +11,7 @@
 
   ============================================================================*/
 
-.dataset-upload-tile {
+.dataset-upload-status {
 	.progress-bar {
 		width: 100%;
 	}
