+++ dataprep-webapp/src/app/components/dataset/upload-status/dataset-upload-status.html
@@ -13,19 +13,19 @@
   ~ ============================================================================
   -->
 
-<div class="dataset-upload-tile inventory-list">
-	<div class="inventory-item" ng-if="uploadTileCtrl.dataset">
+<div class="dataset-upload-status inventory-list">
+	<div class="inventory-item" ng-if="uploadStatusCtrl.dataset">
 		<div class="inventory-icon">
 			<icon name="'talend-file-o'"/>
 		</div>
-		<div class="inventory-text" ng-switch="uploadTileCtrl.dataset.error">
-			<div class="inventory-title">{{uploadTileCtrl.dataset.name}}</div>
+		<div class="inventory-text" ng-switch="uploadStatusCtrl.dataset.error">
+			<div class="inventory-title">{{uploadStatusCtrl.dataset.name}}</div>
 
-			<div class="inventory-progress" ng-switch-when="false" ng-switch="uploadTileCtrl.dataset.type">
+			<div class="inventory-progress" ng-switch-when="false" ng-switch="uploadStatusCtrl.dataset.type">
 				<div class="progress-bar" ng-switch-when="file">
-				<span class="meter" style="width: {{ uploadTileCtrl.dataset.progress }}%" ng-switch="uploadTileCtrl.dataset.progress">
+				<span class="meter" style="width: {{ uploadStatusCtrl.dataset.progress }}%" ng-switch="uploadStatusCtrl.dataset.progress">
 					<p ng-switch-when="100"><span translate-once="UPLOAD_PROCESSING"></span></p>
-					<p ng-switch-default="">{{ uploadTileCtrl.dataset.progress }} %</p>
+					<p ng-switch-default="">{{ uploadStatusCtrl.dataset.progress }} %</p>
 				</span>
 				</div>
 				<div class="progress-bar" ng-switch-when="remote">
