+++ dataprep-webapp/src/app/components/home/home-module.js
@@ -14,7 +14,7 @@
 import angular from 'angular';
 import ABOUT_MODULE from '../about/about-module';
 import DATASET_IMPORT_MODULE from '../dataset/import/dataset-import.module';
-import DATASET_UPDLOAD_TILE_MODULE from '../dataset/upload-tile/dataset-upload-tile-module';
+import DATASET_UPDLOAD_TILE_MODULE from '../dataset/upload-status/dataset-upload-status-module';
 import DATASET_XLS_PREVIEW_MODULE from '../dataset/preview/dataset-xls-preview-module';
 import FOLDER_CREATOR_MODULE from '../folder/creator/folder-creator-module';
 import PREPARATION_CREATOR_MODULE from '../preparation/creator/preparation-creator-module';
@@ -41,7 +41,7 @@ export default MODULE_NAME;
  * @requires talend.about
  * @requires talend.widget
  * @requires data-prep.dataset-import
- * @requires data-prep.dataset-upload-tile
+ * @requires data-prep.dataset-upload-status
  * @requires data-prep.preparation-creator
  * @requires data-prep.preparation-copy-move
  * @requires data-prep.services.dataset
