+++ dataprep-webapp/src/app/components/preparation/creator/form/_preparation-creator-form.scss
@@ -11,7 +11,7 @@
 
   ============================================================================*/
 
-.dataset-upload-tile {
+.dataset-upload-status {
 	max-height: 90px;
 	overflow-y: auto;
 }
