+++ dataprep-webapp/src/app/components/preparation/creator/form/preparation-creator-form.html
@@ -16,10 +16,10 @@
 
 <div class="modal-title" translate-once="ADD_PREPARATION_MODAL_TITLE"></div>
 
-<dataset-upload-tile class="base-dataset-upload"
+<dataset-upload-status class="base-dataset-upload"
 					 id="upload-base-dataset"
 					 dataset="$ctrl.state.dataset.uploadingDataset">
-</dataset-upload-tile>
+</dataset-upload-status>
 
 <div class="preparation-creator-header">
 	<span class="filters-list-title" translate-once="EXISTING_DATASETS"></span>
