+++ dataprep-webapp/src/app/components/widgets-containers/widgets-containers-module.js
@@ -33,7 +33,7 @@ import SidePanelContainer from './side-panel/side-panel-container';
 
 import SETTINGS_MODULE from '../../settings/settings-module';
 import STATE_MODULE from '../../services/state/state-module';
-import DATASET_UPLOAD_TILE_MODULE from '../dataset/upload-tile/dataset-upload-tile-module';
+import DATASET_UPLOAD_STATUS_MODULE from '../dataset/upload-status/dataset-upload-status-module';
 import DATASET_PROGRESS from '../dataset/progress/dataset-progress-module';
 
 const MODULE_NAME = 'react-talend-components.containers';
@@ -44,7 +44,7 @@ angular.module(MODULE_NAME,
 		'pascalprecht.translate',
 		SETTINGS_MODULE,
 		STATE_MODULE,
-		DATASET_UPLOAD_TILE_MODULE,
+		DATASET_UPLOAD_STATUS_MODULE,
 		DATASET_PROGRESS,
 	])
 	.directive('pureAppHeaderBar', ['reactDirective', reactDirective => reactDirective(AppHeaderBar)])
