+++ dataprep-webapp/src/app/index.scss
@@ -45,7 +45,7 @@ $sc-dropdown-content-color: initial;
 @import "components/dataset/progress/_dataset-progress.scss";
 @import "components/dataset/parameters/_dataset-parameters.scss";
 @import "components/dataset/preview/_dataset-xls-preview.scss";
-@import "components/dataset/upload-tile/_dataset-upload-tile.scss";
+@import "components/dataset/upload-status/_dataset-upload-status.scss";
 @import "components/easter-eggs/star-wars/_star-wars.scss";
 @import "components/filter/bar/_filter-bar.scss";
 @import "components/filter/item/_filter-item.scss";
